+++ Helpers/OptimizationOptions.cs
@@ -466,9 +466,6 @@ internal partial class OptimizationOptions
                     case "EnableRecommendedSectionStartMenu":
                         OptimizeSystemHelper.EnableRecommendedSectionStartMenu();
                         break;
-                    case "EnableUnnecessaryServices":
-                        OptimizeSystemHelper.EnableUnnecessaryServices();
-                        break;
                     case "DisableLegacyBootMenu":
                         OptimizeSystemHelper.DisableLegacyBootMenu();
                         break;
@@ -745,6 +742,9 @@ internal partial class OptimizationOptions
                     case "EnableCoPilotAI":
                         OptimizeSystemHelper.EnableCoPilotAI();
                         break;
+                    case "EnableWindowsRecall":
+                        OptimizeSystemHelper.EnableWindowsRecall();
+                        break;
                     case "EnableVisualStudioTelemetry":
                         OptimizeSystemHelper.EnableVisualStudioTelemetry();
                         break;
@@ -796,19 +796,6 @@ internal partial class OptimizationOptions
                     {
                         OptimizeSystemHelper.DisableLegacyBootMenu();
                         RemoveRevertAction("DisableLegacyBootMenu");
-                    }
-                    break;
-
-                case "UnnecessaryServices":
-                    if (toggleSwitch.IsOn)
-                    {
-                        OptimizeSystemHelper.DisableUnnecessaryServices();
-                        SaveRevertAction("EnableUnnecessaryServices");
-                    }
-                    else
-                    {
-                        OptimizeSystemHelper.EnableUnnecessaryServices();
-                        RemoveRevertAction("EnableUnnecessaryServices");
                     }
                     break;
 
@@ -2023,6 +2010,19 @@ internal partial class OptimizationOptions
                     {
                         OptimizeSystemHelper.EnableCoPilotAI();
                         RemoveRevertAction("EnableCoPilotAI");
+                    }
+                    break;
+
+                case "WindowsRecall":
+                    if (toggleSwitch.IsOn)
+                    {
+                        OptimizeSystemHelper.DisableWindowsRecall();
+                        SaveRevertAction("EnableWindowsRecall");
+                    }
+                    else
+                    {
+                        OptimizeSystemHelper.EnableWindowsRecall();
+                        RemoveRevertAction("EnableWindowsRecall");
                     }
                     break;